refactor(20-08w6): update data types and foreign key structures
- Change data type from double to decimal for precision
- Replace string class code with foreign key 'int lopHocId'
- Replace string department code with foreign key 'int BoMonId'
- Update endpoint for getting class list to match new structure

+++ Migrations/20260817094628_createNew.cs
@@ -56,7 +56,7 @@ namespace StudentAPIw6.Migrations
                     GioiTinh = table.Column<bool>(type: "bit", nullable: false),
                     NgaySinh = table.Column<DateTime>(type: "datetime2", nullable: false),
                     Email = table.Column<string>(type: "nvarchar(max)", nullable: false),
-                    DiemTB = table.Column<double>(type: "float", nullable: false),
+                    DiemTB = table.Column<decimal>(type: "float", nullable: false),
                     MaLop = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     lopHocId = table.Column<string>(type: "nvarchar(450)", nullable: true)
                 },
